requirements.txt
fastapi==0.115.8
pydantic==2.10.6
ollama==0.4.7
pyaml==25.1.0
uvicorn==0.34.0
python-dotenv==1.0.1